requirements.txt
notion-client>=2.2.1
python-dotenv>=1.0.0
pydantic>=2.5.2
fastapi>=0.109.0
uvicorn>=0.27.0
numpy==1.24.3
scikit-learn==1.2.2
pandas==1.5.3
chromadb>=0.4.14
sentence-transformers>=2.5.1
requests>=2.31.0
transformers>=4.36.2
pytest>=7.4.0
pytest-cov>=4.1.0 